+++ src/main/java/com/demo/hotel/HotelBackendApplication.java
@@ -1,4 +1,4 @@
-package com.demo.hotelbackend;
+package com.demo.hotel;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
