+++ src/test/java/com/demo/hotel/HotelBackendApplicationTests.java
@@ -1,4 +1,4 @@
-package com.demo.hotelbackend;
+package com.demo.hotel;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
